requirements.txt
django==4.2.7
djangorestframework==3.15.2
drf-spectacular==0.28.0
psycopg2==2.9.9
python-decouple==3.8
requests==2.32.3
